+++ Encoding/x264Enc.vb
@@ -52,8 +52,9 @@ Public Class x264Enc
         If Params.Mode.Value = x264RateMode.TwoPass Then
             Encode("Video encoding second pass", GetArgs(2, p.Script), s.ProcessPriority)
         ElseIf Params.Mode.Value = x264RateMode.ThreePass Then
-            Encode("Video encoding second pass", GetArgs(3, p.Script), s.ProcessPriority)
-            Encode("Video encoding third pass", GetArgs(2, p.Script), s.ProcessPriority)
+            'Specific order 1 > 3 > 2 is correct!
+            Encode("Video encoding Nth pass", GetArgs(3, p.Script), s.ProcessPriority)
+            Encode("Video encoding last pass", GetArgs(2, p.Script), s.ProcessPriority)
         End If
 
         AfterEncoding()
@@ -1077,6 +1078,10 @@ Public Class x264Params
             args += " " + q.Select(Function(item) item.GetArgs).Join(" ")
         End If
 
+        If args.Contains("%") Then
+            args = Macro.Expand(args)
+        End If
+
         If includePaths Then
             Dim input = If(pipeTool = "none", script.Path.ToShortFilePath.Escape, "-")
             Dim dmx = Demuxer.ValueText
@@ -1093,7 +1098,6 @@ Public Class x264Params
 
             If dmx <> "" Then
                 Dim info = script.GetInfo
-
                 args += $" --demuxer {dmx} --frames " & info.FrameCount
 
                 If dmx = "raw" Then
@@ -1118,7 +1122,7 @@ Public Class x264Params
             End If
         End If
 
-        Return Macro.Expand(args.Trim.FixBreak.Replace(BR, " "))
+        Return args.Trim.FixBreak.Replace(BR, " ")
     End Function
 
     Function GetPartitionsArg() As String